modules/location-producer/requirements.txt
grpcio==1.50.0
protobuf==4.21.8
grpcio-tools==1.50.0
kafka-python==2.0.2
